+++ changelog.md
@@ -19,7 +19,7 @@
 * Updated kh mobile icons
 * Added more suspend time options in the server admin page
 * Removed deprecated scripts
-* Added script to install kamehouse-shell scripts standalone
+* Added script to install kamehouse-shell scripts only
 * Updated the installation and environment setup guides and other MDs
 * Added install script to automate most of kamehouse installation process
 * Added uninstall script to uninstall kamehouse and purge the config files